Integrate refactored DaL regressor
