--- wluncert/DaL-ext-sklearn-rework/requirements.txt
@@ -1,13 +1,0 @@
-doepy~=0.0.1
-#imbalanced_learn==0.12.0
-#imblearn==0.0
-matplotlib==3.6.2
-numpy==1.24.3
-scipy<=1.11.0
-pandas~=1.3.1
-pymoo==0.6.0.1
-#scikit_learn==1.4.2
-scikit-learn~=1.2.2
-imbalanced-learn~=0.11.0
-
-tensorflow==2.15.0
