+++ wluncert/dal/__init__.py
@@ -1,0 +1,1 @@
+from .regressor import DaLRegressor
