+++ wluncert/dal/utils/general.py
